Rename interface of loading spinner
