+++ src/components/UI/LoadingSpinner/LoadingSpinner.interface.ts
